@@ -1,40 +1,43 @@
 <?php
 
+declare(strict_types=1);
+
 namespace Autowp\ZFComponents\Db;
 
-use Zend\Db\Adapter\Adapter;
-use Zend\Db\TableGateway\Feature\SequenceFeature;
-use Zend\Db\TableGateway\TableGateway;
-use Zend\ServiceManager\Exception\ServiceNotFoundException;
-use Zend\ServiceManager\ServiceLocatorInterface;
+use Laminas\Db\Adapter\Adapter;
+use Laminas\Db\TableGateway\Feature\SequenceFeature;
+use Laminas\Db\TableGateway\TableGateway;
+use Laminas\ServiceManager\Exception\ServiceNotFoundException;
+use Laminas\ServiceManager\ServiceLocatorInterface;
+use Psr\Container\ContainerExceptionInterface;
+use Psr\Container\NotFoundExceptionInterface;
+
+use function sprintf;
 
 class TableManager implements ServiceLocatorInterface
 {
-    /**
-     * @var array
-     */
-    private $specs = [];
+    /** @var array */
+    private array $specs = [];
 
-    /**
-     * @var array
-     */
-    private $tables = [];
+    /** @var array */
+    private array $tables = [];
 
-    /**
-     * @var Adapter
-     */
-    private $adapter;
+    /** @var Adapter */
+    private Adapter $adapter;
 
     public function __construct(Adapter $adapter, array $specs)
     {
         $this->adapter = $adapter;
-        $this->specs = $specs;
+        $this->specs   = $specs;
     }
 
     /**
      * @SuppressWarnings(PHPMD.UnusedFormalParameter)
+     * @param  string $name
+     * @return TableGateway
+     * @throws ContainerExceptionInterface If any other error occurs.
      */
-    public function build($name, array $options = null)
+    public function build($name, ?array $options = null)
     {
         if (! isset($this->specs[$name])) {
             throw new ServiceNotFoundException(sprintf(
@@ -45,11 +48,11 @@ class TableManager implements ServiceLocatorInterface
 
         $spec = $this->specs[$name];
 
-        $platform = $this->adapter->getPlatform();
+        $platform     = $this->adapter->getPlatform();
         $platformName = $platform->getName();
 
         $features = [];
-        if ($platformName == 'PostgreSQL' && isset($spec['sequences'])) {
+        if ($platformName === 'PostgreSQL' && isset($spec['sequences'])) {
             foreach ($spec['sequences'] as $field => $sequence) {
                 $features[] = new SequenceFeature($field, $sequence);
             }
@@ -58,6 +61,12 @@ class TableManager implements ServiceLocatorInterface
         return new TableGateway($name, $this->adapter, $features);
     }
 
+    /**
+     * @param string $id Identifier of the entry to look for.
+     * @throws NotFoundExceptionInterface  No entry was found for **this** identifier.
+     * @throws ContainerExceptionInterface Error while retrieving the entry.
+     * @return mixed Entry.
+     */
     public function get($id)
     {
         if (! isset($this->tables[$id])) {
@@ -67,6 +76,10 @@ class TableManager implements ServiceLocatorInterface
         return $this->tables[$id];
     }
 
+    /**
+     * @param string $id Identifier of the entry to look for.
+     * @return bool
+     */
     public function has($id)
     {
         return isset($this->specs[$id]);
