@@ -1,19 +1,21 @@
 <?php
 
-namespace Autowp\ZFComponents\Factory;
+declare(strict_types=1);
 
-use Interop\Container\ContainerInterface;
-use Zend\ServiceManager\Factory\FactoryInterface;
+namespace Autowp\ZFComponents\Factory;
 
 use Autowp\ZFComponents\GulpRev as Service;
 use Autowp\ZFComponents\View\Helper\GulpRev as Helper;
+use Interop\Container\ContainerInterface;
+use Laminas\ServiceManager\Factory\FactoryInterface;
 
 class GulpRevViewHelperFactory implements FactoryInterface
 {
     /**
+     * @param  string $requestedName
      * @SuppressWarnings(PHPMD.UnusedFormalParameter)
      */
-    public function __invoke(ContainerInterface $container, $requestedName, array $options = null)
+    public function __invoke(ContainerInterface $container, $requestedName, ?array $options = null): Helper
     {
         $service = $container->get(Service::class);
 
