@@ -1,10 +1,11 @@
 <?php
 
+declare(strict_types=1);
+
 namespace Autowp\ZFComponents;
 
 use Exception;
 
 class GulpRevException extends Exception
 {
-
 }
