@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 namespace Autowp\ZFComponents;
 
 class Module
@@ -13,7 +15,7 @@ class Module
             'view_helpers'    => $provider->getViewHelperConfig(),
             'service_manager' => $provider->getDependencyConfig(),
             'tables'          => [],
-            'rollbar'         => $provider->getRollbarConfig()
+            'rollbar'         => $provider->getRollbarConfig(),
         ];
     }
 }
