@@ -1,20 +1,20 @@
 <?php
 
+declare(strict_types=1);
+
 namespace AutowpTest\ZFComponents;
 
-use Zend\Db\TableGateway\TableGateway;
-use Zend\Mvc\Application;
-use Zend\ServiceManager\ServiceManager;
+use Laminas\Db\TableGateway\TableGateway;
+use Laminas\Mvc\Application;
+use PHPUnit\Framework\TestCase;
 
-use Autowp\ZFComponents\Db\TableManager;
-
-class DbTableManagerTest extends \PHPUnit\Framework\TestCase
+class DbTableManagerTest extends TestCase
 {
-    public function testGetTable()
+    public function testGetTable(): void
     {
-        $app = Application::init(require __DIR__ . '/_files/config/application.config.php');
+        $app            = Application::init(require __DIR__ . '/_files/config/application.config.php');
         $serviceManager = $app->getServiceManager();
-        $tables = $serviceManager->get('TableManager');
+        $tables         = $serviceManager->get('TableManager');
 
         $table = $tables->get('foo');
 
