@@ -1,122 +1,130 @@
 <?php
 
+declare(strict_types=1);
+
 namespace AutowpTest\ZFComponents;
 
-class WildcardSafeRouteTest extends \PHPUnit\Framework\TestCase
-{
-    private $app = null;
+use Laminas\Http\Request;
+use Laminas\Mvc\Application;
+use Laminas\Router\RouteMatch;
+use Laminas\View\Renderer\RendererInterface;
+use PHPUnit\Framework\TestCase;
 
-    private function getApp()
+class WildcardSafeRouteTest extends TestCase
+{
+    private $app;
+
+    private function getApp(): Application
     {
         if ($this->app === null) {
-            $this->app = \Zend\Mvc\Application::init(require __DIR__ . '/_files/config/application.config.php');
+            $this->app = Application::init(require __DIR__ . '/_files/config/application.config.php');
         }
 
         return $this->app;
     }
 
-    private function getView()
+    private function getView(): RendererInterface
     {
         $serviceManager = $this->getApp()->getServiceManager();
 
         return $serviceManager->get('ViewRenderer');
     }
 
-    public function testAssemble()
+    public function testAssemble(): void
     {
         $url = $this->getView()->url('example/params', [
             'param1' => 'value1',
-            'param2' => 'value2'
+            'param2' => 'value2',
         ]);
 
         $this->assertEquals('/example/param1/value1/param2/value2', $url);
     }
 
-    public function testMatch()
+    public function testMatch(): void
     {
         $serviceManager = $this->getApp()->getServiceManager();
-        $router = $serviceManager->get('HttpRouter');
+        $router         = $serviceManager->get('HttpRouter');
 
-        $request = \Zend\Http\Request::fromString("GET /example/param1/value1/param2/value2 HTTP/1.0\n");
+        $request = Request::fromString("GET /example/param1/value1/param2/value2 HTTP/1.0\n");
 
         $match = $router->match($request);
 
         $this->assertNotNull($match);
-        $this->assertInstanceOf(\Zend\Router\RouteMatch::class, $match);
+        $this->assertInstanceOf(RouteMatch::class, $match);
 
         $this->assertEquals('example/params', $match->getMatchedRouteName());
         $this->assertEquals([
             'param1' => 'value1',
-            'param2' => 'value2'
+            'param2' => 'value2',
         ], $match->getParams());
     }
 
-    public function testAssembleWithEmptyParams()
+    public function testAssembleWithEmptyParams(): void
     {
         $url = $this->getView()->url('example/params', []);
 
         $this->assertEquals('/example', $url);
     }
 
-    public function testAssembleWithCustomDelimiter()
+    public function testAssembleWithCustomDelimiter(): void
     {
         $url = $this->getView()->url('example2/params', [
             'param1' => 'value1',
-            'param2' => 'value2'
+            'param2' => 'value2',
         ]);
 
         $this->assertEquals('/example2/param1=value1/param2=value2', $url);
     }
 
-    public function testMatchWithCustomDelimiter()
+    public function testMatchWithCustomDelimiter(): void
     {
         $serviceManager = $this->getApp()->getServiceManager();
-        $router = $serviceManager->get('HttpRouter');
+        $router         = $serviceManager->get('HttpRouter');
 
-        $request = \Zend\Http\Request::fromString("GET /example2/param1=value1/param2=value2 HTTP/1.0\n");
+        $request = Request::fromString("GET /example2/param1=value1/param2=value2 HTTP/1.0\n");
 
         $match = $router->match($request);
 
         $this->assertNotNull($match);
-        $this->assertInstanceOf(\Zend\Router\RouteMatch::class, $match);
+        $this->assertInstanceOf(RouteMatch::class, $match);
 
         $this->assertEquals('example2/params', $match->getMatchedRouteName());
         $this->assertEquals([
             'param1' => 'value1',
-            'param2' => 'value2'
+            'param2' => 'value2',
         ], $match->getParams());
     }
 
-    public function testAssembleWithExcludedParameters()
+    public function testAssembleWithExcludedParameters(): void
     {
         $url = $this->getView()->url('example/params', [
             'param1'     => 'value1',
             'param2'     => 'value2',
             'controller' => 'test',
-            'action'     => 'test'
+            'action'     => 'test',
         ]);
 
         $this->assertEquals('/example/param1/value1/param2/value2', $url);
     }
 
-    public function testMatchWithExcludedParameters()
+    public function testMatchWithExcludedParameters(): void
     {
         $serviceManager = $this->getApp()->getServiceManager();
-        $router = $serviceManager->get('HttpRouter');
+        $router         = $serviceManager->get('HttpRouter');
 
-        $request = \Zend\Http\Request::fromString(
+        $request = Request::fromString(
             "GET /example/param1/value1/param2/value2/controller/test/action/test HTTP/1.0\n"
         );
 
         $match = $router->match($request);
 
         $this->assertNotNull($match);
-        $this->assertInstanceOf(\Zend\Router\RouteMatch::class, $match);
+        $this->assertInstanceOf(RouteMatch::class, $match);
 
         $this->assertEquals('example/params', $match->getMatchedRouteName());
         $this->assertEquals([
             'param1' => 'value1',
-            'param2' => 'value2'
+            'param2' => 'value2',
         ], $match->getParams());
     }
 }
