@@ -1,19 +1,21 @@
 <?php
 
+declare(strict_types=1);
+
 return [
-    'modules' => [
-        'Zend\\Db',
-        'Zend\\Filter',
-        'Zend\\Router',
-        'Zend\\I18n',
-        'Autowp\\ZFComponents'
+    'modules'                 => [
+        'Laminas\\Db',
+        'Laminas\\Filter',
+        'Laminas\\Router',
+        'Laminas\\I18n',
+        'Autowp\\ZFComponents',
     ],
     'module_listener_options' => [
-        'module_paths' => [
+        'module_paths'      => [
             './vendor',
         ],
         'config_glob_paths' => [
             'test/_files/config/autoload/local.php',
         ],
-    ]
+    ],
 ];
